% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeDuplicateCount.R
\name{cumulativeDuplicateCount}
\alias{cumulativeDuplicateCount}
\title{Get the cumulative duplicate count}
\usage{
cumulativeDuplicateCount(x)
}
\arguments{
\item{x}{A vector with (potential) duplicates}
}
\value{
A vector of length(x) with the cumulative duplicate count
}
\description{
Get the cumulative duplicate count
}
\examples{
metabefor::cumulativeDuplicateCount(
  c(1, 1, 2, 3, 4, 2, 5, 6, 2, 7, 8, 1, 2)
);
}
