% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_object_to_yaml.R
\name{dct_object_to_yaml}
\alias{dct_object_to_yaml}
\title{Convert a DCT object to YAML}
\usage{
dct_object_to_yaml(dctObject)
}
\arguments{
\item{dctObject}{The DCT object}
}
\value{
A character vector.
}
\description{
Convert a DCT object to YAML
}
