% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_sheet_to_dct.R
\name{dct_sheet_to_dct}
\alias{dct_sheet_to_dct}
\title{Create a DCT object from a DCT sheet}
\usage{
dct_sheet_to_dct(dct_sheet)
}
\arguments{
\item{dct_sheet}{A dataframe containing a DCT specification.}
}
\value{
A DCT created by \code{\link[=dct_object]{dct_object()}}.
}
\description{
Create a DCT object from a DCT sheet
}
