% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_doi.R
\name{extract_doi}
\alias{extract_doi}
\title{Extract DOI(s) from a character vector}
\usage{
extract_doi(x)
}
\arguments{
\item{x}{The character vector.}
}
\value{
A character vector.
}
\description{
This function extracts the DOI or DOIs from a character vector. It always
returns a single DOI for each element in the input vector (if one is found),
but returns additional DOIs in the attributes.
}
\examples{
exampleVector <-
  c("eni040 [pii] and 10.1093/deafed/eni040 [doi]",
    "S0022437503000938 [pii] and 10.1016/j.jsr.2003.12.001 [doi]",
    NA,
    "10.1080/15428119791012261 [doi]");
metabefor::extract_doi(
  exampleVector
);
}
