% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenNodeValue.R, R/flattenNodeValues.R
\name{flattenNodeValue}
\alias{flattenNodeValue}
\alias{flattenNodeValues}
\title{Flatten one or more node values}
\usage{
flattenNodeValue(singleValue)

flattenNodeValues(x)
}
\arguments{
\item{singleValue}{The single node value}

\item{x}{A list of node values}
}
\value{
A character value, or a vector of values.
}
\description{
"Flattening" a ndoe value means pasting together all values in the vector
to one character value.
}
\examples{
metabefor::flattenNodeValue(1:8);

metabefor::flattenNodeValues(
  list(
    1:8,
    letters[1:4],
    LETTERS[10:17]
  )
);
}
