% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.odd_is.even.R
\name{is.odd}
\alias{is.odd}
\alias{is.even}
\title{Checking whether numbers are odd or even}
\usage{
is.odd(vector)

is.even(vector)
}
\arguments{
\item{vector}{The vector to process}
}
\value{
A logical vector.
}
\description{
Checking whether numbers are odd or even
}
\examples{
is.odd(4);
}
