% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeMatrices.R, R/maximizeMatrix.R
\name{maximizeMatrices}
\alias{maximizeMatrices}
\alias{maximizeMatrix}
\title{Adding columns or rows to a matrix}
\usage{
maximizeMatrices(x, fillValue = 0)

maximizeMatrix(x, rowNames, colNames, fillValue = 0)
}
\arguments{
\item{x}{Either a single data frame or matrix (for \code{maximizeMatrix}) or,
more commonly, a list of data frames or matrces (for \code{maximizeMatrices}).}

\item{fillValue}{The value to fill the new rows and columns with}

\item{rowNames, colNames}{The names or the rows and columns as they
should be in the final result}
}
\value{
A matrix with the specified rows and columns
}
\description{
Created to add missing rows or columns to a matrix
}
\examples{
metabefor::maximizeMatrix(
  mtcars[1:3, 1:4],
  rowNames = c('Mazda RX4', letters[1:3]),
  colNames = c('mpg', 'cyl', LETTERS[1:4])
);
}
