% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rxs_parsedExtractionScripts.R
\name{print.rxs_parsedExtractionScripts}
\alias{print.rxs_parsedExtractionScripts}
\title{Printing a project with parsed Rxs file}
\usage{
\method{print}{rxs_parsedExtractionScripts}(x, ...)
}
\arguments{
\item{x}{The Rxs project file (as produced
by \code{\link[=rxs_parseExtractionScripts]{rxs_parseExtractionScripts()}})}

\item{...}{Any additional arguments (ignored)}
}
\value{
\code{x}, invisibly.
}
\description{
Printing a project with parsed Rxs file
}
