% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rxsStructure.R
\name{print.rxsStructure}
\alias{print.rxsStructure}
\title{Printing a parsed Rxs specification}
\usage{
\method{print}{rxsStructure}(x, ...)
}
\arguments{
\item{x}{The Rxs specification (as produced
by \code{\link[=rxs_fromSpecifications]{rxs_fromSpecifications()}})}

\item{...}{Any additional arguments (ignored)}
}
\value{
\code{x}, invisibly.
}
\description{
Printing a parsed Rxs specification
}
