% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_df_list.R
\name{rbind_df_list}
\alias{rbind_df_list}
\title{Bind lots of dataframes together rowwise}
\usage{
rbind_df_list(x, suppressNonDfError = FALSE)
}
\arguments{
\item{x}{A list of dataframes}

\item{suppressNonDfError}{Whether to suppress the error when the object
that is passed is not a data frame.}
}
\value{
A dataframe
}
\description{
Bind lots of dataframes together rowwise
}
\examples{
rbind_df_list(list(Orange, mtcars, ChickWeight));
}
