% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json_from_url.R
\name{read_json_from_url}
\alias{read_json_from_url}
\title{Read an object from JSON embedded in a HTML page}
\usage{
read_json_from_url(url, attribute, elementId = NULL, elementClass = NULL)
}
\arguments{
\item{url}{The URL where the JSON can be found.}

\item{attribute}{The attribute holding the JSON.}

\item{elementId, elementClass}{Specify exactly one of these: either the
identifier of the element holding the JSON, or the class name of a class
that may be applied to several elements (which will then all be returned).}
}
\value{
An object if specifying \code{elementId}, or a list of objects if
specifying \code{elementClass} (although is there's only one element of
\code{elementClass}, the only object will returned anyway).
}
\description{
You normally don't use this; this function is exposed for convenience,
but you normally use functions that import specific things (e.g.
questionnaires).
}
