% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SOQs_from_url.R
\name{read_SOQs_from_url}
\alias{read_SOQs_from_url}
\title{Read all SOQs from a URL}
\usage{
read_SOQs_from_url(
  url = "https://operationalizations.com/questionnaires/json",
  elementClass = "operationalizations-com-json",
  attribute = "data-operationalizations-com"
)
}
\arguments{
\item{url}{The URL to the repo}
}
\value{
An object of imported SOQs
}
\description{
This function imports all Serialized Open Questionnaire
specifications (i.e. SOQ specs) from an online repository.
}
\examples{
\donttest{
all_soq_specs <-
  psyverse::read_SOQs_from_url();
}
}
