% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included_data_resources.R
\docType{data}
\name{rxs_minimal_example_2}
\alias{rxs_minimal_example_2}
\title{A very rudimentary example Rxs specification}
\format{
An example of an Rxs specification.
}
\usage{
rxs_minimal_example_2
}
\description{
This is a simple and relatively short Rxs specification.
}
\keyword{datasets}
