% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxs_parseSpecifications.R
\name{rxs_parseSpecifications}
\alias{rxs_parseSpecifications}
\title{Parse Rxs (R extraction script) specifications}
\usage{
rxs_parseSpecifications(
  entities,
  valueTemplates,
  definitions = NULL,
  moduleName = NULL,
  rxsRootName = metabefor::opts$get("rxsRootName"),
  silent = metabefor::opts$get("silent")
)
}
\arguments{
\item{entities}{The entity specification, as a data frame}

\item{valueTemplates}{The value templates, as a data frame}

\item{definitions}{The definitions, as a data frame}

\item{moduleName}{The name of the module; only used to make error messages
more informative, and only used if not NULL.}

\item{rxsRootName}{The name of the root node of the Rxs tree}

\item{silent}{Whether to be silent or chatty.}
}
\value{
An \code{rxsStructure} object
}
\description{
This function parses the specifications from
an Rxs specification spreadsheet.
}
