% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsTree_has_entity.R
\name{rxsTree_has_entity}
\alias{rxsTree_has_entity}
\title{Check whether a Rxs tree has an entity}
\usage{
rxsTree_has_entity(x, entityId_regex)
}
\arguments{
\item{x}{The Rxs tree.}

\item{entityId_regex}{The regular expression}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check whether a Rxs tree has an entity with an identifier that
matches the regular expression.
}
