% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_search.R
\name{specify_search}
\alias{specify_search}
\title{Specify a search}
\usage{
specify_search(date, query, databases, fields)
}
\arguments{
\item{date}{The date that the search is conducted.}

\item{query}{The query used in the search (as specified using the
functions for queries, e.g. \code{\link[=query_full]{query_full()}}).}

\item{databases}{The databases that were searched, as a named list of
vectors, where each vector's name is the name of the interface used to
access the databases in the vector.}

\item{fields}{The fields in which the query terms were searched.}
}
\value{
An \code{mbf_search} object with the search specifications.
}
\description{
Specify a search
}
