% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toi_to_soi.R
\name{toi_to_soi}
\alias{toi_to_soi}
\title{Import a TOI spec and convert it to a SOI spec}
\usage{
toi_to_soi(x, returnYAML = FALSE)
}
\arguments{
\item{x}{The path to a spreadheet file or the URL to a google sheet that
is publicly accessible.}

\item{returnYAML}{Whether to return an R object (\code{FALSE}) or YAML (\code{TRUE}).}
}
\value{
A character vector holding the YAML.
}
\description{
Import a Tabulated Open Item specification (e.g. in a
spreadsheet) and convert it into a Serialized Open Item
specification (as an R object or in YAML format).
}
\examples{
\donttest{
soi <-
  psyverse::toi_to_soi(
    paste0(
      "https://docs.google.com/spreadsheets/d/",
      "1T7UN7tiLV89YXMglXdC7Cu3nCAijt_1mV93ehwM3ybE"
    ),
    returnYAML = TRUE
  );
}

}
