% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewHTML.R
\name{viewHTML}
\alias{viewHTML}
\title{Display HTML}
\usage{
viewHTML(x, title = "Psyverse", css = "body {font-size: 16px;}")
}
\arguments{
\item{x}{The HTML fragment}

\item{title}{The title}

\item{css}{CSS}
}
\value{
Invisibly, \code{x}, with the extra HTML bits added.
}
\description{
This function displays HTML in the viewer, adding \verb{<body>} and \verb{<head>}
tags (which should therefore not be included in the fragment).
}
\examples{
### Only run this example in an interactive R session,
### as it shows the HTML in the viewer.
if (interactive()) {
  psyverse::viewHTML("<strong>Hello world!</strong>");
}
}
