% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_error.R
\name{wrap_error}
\alias{wrap_error}
\title{Wrap an error message to try to get users to read it}
\usage{
wrap_error(...)
}
\arguments{
\item{...}{Character values that are concatenated without separator.}
}
\value{
The wrapped message.
}
\description{
Pretty much does what it says on the tin.
}
\examples{
cat(
  metabefor::wrap_error(
    "You missed a closing quote somewhere!"
  )
);
}
